Embed map swipe: moved to lib
