+++ lib/leaflet-side-by-side/index.js
