+++ lib/leaflet-side-by-side/range-icon.png
